+++ R/swirlify.R
@@ -1,5 +1,4 @@
 # Creates skeleton for new course/lesson.
-#' @importFrom utils packageVersion
 make_skeleton <- function() {
   # Course directory name
   courseDirName <- gsub(" ", "_", getOption("swirlify_course_name"))
@@ -36,65 +35,81 @@ make_skeleton <- function() {
   return(lessonPath)
 }
 
-# Save the lesson YAML to options.
-# 
-# This function is for internal use only, but is made public
-# so that it is accessible from the content authoring app.
-# 
-# @param lesson lesson name
-# @param course course name
-# @param author your name
-# @param organization your organization's name
-# @importFrom shinyAce aceEditor updateAceEditor
-# @export
-# setMeta <- function(lesson, course, author, organization) {
-#   if(is.null(lesson)) lesson <- "Lesson name here"
-#   if(is.null(course)) course <- "Course name here"
-#   if(is.null(author)) author <- "Your name here"
-#   if(is.null(organization)) organization <- "Your organization name here (optional)"
-#   options(swirlify_lesson_name = lesson,
-#           swirlify_course_name = course,
-#           swirlify_author = author,
-#           swirlify_organization = organization)
-#   invisible()
-# }
+#' Lauch a Shiny application for writing swirl lessons
+#' 
+#' This function launches a user interface for writing
+#' swirl lessons.
+#' 
+#' @param lesson_name The name of the new lesson you want to
+#' create. The default value is \code{NULL}. If you've
+#' already selected a lesson to work on using \code{\link{set_lesson}}
+#' then you do not need to provide a value for this argument.
+#' @param course_name The name of the new course you want to
+#' create. The default value is \code{NULL}. If you've
+#' already selected a course to work on using \code{\link{set_lesson}}
+#' then you do not need to provide a value for this argument.
+#' 
+#'
+#' @import shiny
+#' @import swirl
+#' @import shinyAce
+#' @export
+#' @examples 
+#' \dontrun{
+#' 
+#' # Set lesson beforehand
+#' set_lesson()
+#' swirlify()
+#' 
+#' # Start a new lesson in your current directory
+#' swirlify("Lesson 1", "My Course")
+#' 
+#' }
+swirlify <- function(lesson_name = NULL, course_name = NULL){
+  if(is.null(lesson_name) && is.null(course_name)){
+    lesson_fp <- getOption("swirlify_lesson_file_path")
+    if(!is.null(lesson_fp) && file.exists(lesson_fp)){
+      startApp()
+    } else {
+      stop("Swirlify cannot find the lesson you are trying to work on. ", 
+           "Please provide arguments for both lesson_name and course_name to start a new lesson, ",
+           "or choose a lesson to work on using set_lesson.")
+    }
+  } else if(!is.null(lesson_name) && !is.null(course_name)){
+    lesson <- get_lesson(lesson_name, course_name)
+    if(is.na(lesson)){
+      new_lesson(lesson_name, course_name, open_lesson = FALSE)
+    } else {
+      set_lesson(lesson, open_lesson = FALSE, silent = TRUE)
+    }
+    startApp()
+  } else {
+    stop("Please provide arguments for both lesson_name and course_name to start a new course.")
+  }
+  invisible()
+}
 
-# Author lesson using content authoring app.
-# 
-# @param lesson lesson name
-# @param course course name
-# @param author your name
-# @param organization your organization's name
-# 
-# @examples
-# \dontrun{
-# 
-# swirlify("Linear Regression", Statistics 101")
-# }
-# @import shiny
-# @import swirl
-# @importFrom methods loadMethod
-# @export
-# swirlify <- function(lesson, course, author=NULL, organization=NULL) {
-#   # Save course/lesson metadata to options for in-app access
-#   setMeta(lesson, course, author, organization)
-#   # Create course skeleton
-#   lessonPath <- make_skeleton()
-#   # Save lesson path to options
-#   options(swirlify_lesson_path = lessonPath)
-#   # Run authoring app
-#   x <- runApp(system.file("fullapp", package="swirlify"))
-#   # Write lesson to file
-#   message("\nWriting lesson to ", lessonPath, " ...")  
-#   writeLines(x[[1]], lessonPath)
-#   
-#   if(isTRUE(x$test)) {
-#     # Course directory name
-#     courseDirName <- gsub(" ", "_", course)
-#     # Install course
-#     install_course_directory(courseDirName)
-#     # Run lesson in "test" mode
-#     swirl("test", test_course=course, test_lesson=lesson)
-#   }
-#   invisible()
-# }
+startApp <- function(){
+  appDir <- system.file("swirlify-app", package = "swirlify")
+  
+  app_results <- list()
+  app_results <- runApp(appDir, display.mode = "normal")
+  
+  if(isTRUE(app_results$demo)){
+    course_path <- getOption("swirlify_course_dir_path")
+    course_name <- getOption("swirlify_course_name")
+    lesson_name <- getOption("swirlify_lesson_name")
+    install_course_directory(course_path)
+    app_results$demo_num <- ifelse(app_results$demo_num < 1, 1, app_results$demo_num)
+    swirl("test", test_course = course_name, test_lesson = lesson_name, 
+          from = app_results$demo_num)
+  }
+  invisible()
+}
+
+get_lesson <- function(lesson_name, course_name){
+  lesson_name <- make_pathname(lesson_name)
+  course_name <- make_pathname(course_name)
+  lessons <- file.path(getwd(), course_name, lesson_name, c("lesson.yaml", "lesson"))
+  Filter(file.exists, lessons)[1]
+}